**feat: add float16 parsing support and improve test/assert coverage**
- Updated `.gitignore` to include profiling and coverage artifacts.
- Replaced `uint8_t` with `float16_t` for GLTF asset versioning.
- Improved `GLTF_t$bind` to support flexible bind formats with `GLTF_BIND_PAIR_SIZE`.
- Removed unused `convertFloat16ToDecimal` function in favor of direct float16 parsing.
- Refactored and simplified `concatBeforeAndAfterString` for clarity and performance.
- Added `ASSERT_TRUE` and `ASSERT_FALSE` macros for easier test assertions.
- Introduced new unit tests for `stringToBool`, `contains`, and `containsString`.
- Updated build flags for profiling with `gcov` and improved debug info with `-ggdb`.

+++ GLTF_t/include/GLTF_asset.h
@@ -2,11 +2,13 @@
 
 #include <stdint.h>
 
+#include "stdfloat16.h"
+
 #define DEFAULT_GLTF_ASSET \
     { .version = 0, .generator = NULL, .copyright = NULL }
 
 struct GLTF_asset {
-    uint8_t version;
+    float16_t version;
     char* generator;
     char* copyright;
 };
